Add move class for representing moves

+++ Loki/position/square.hpp
@@ -6,7 +6,7 @@
 
 namespace loki::position
 {
-	enum e_square : uint64_t
+	enum e_square : uint8_t
 	{
 		A1, B1, C1, D1, E1, F1, G1, H1,
 		A2, B2, C2, D2, E2, F2, G2, H2,
